pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "investment-toolkit"
version = "1.0.0"
description = "Investment analysis toolkit for equity markets"
requires-python = ">=3.14"
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.32.0",
    "yfinance>=0.2.40",
    "fredapi>=0.5.0",
    "plotly>=5.24.0",
    "matplotlib>=3.9.0",
    "TA-Lib>=0.4.32",
    "pushover_complete>=1.1.0",
    "httpx>=0.28.0",
    "aiohttp>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
investment_toolkit = ["py.typed"]
